Initialize OverriddenBaseMethod for imported methods
Due to an unfortunate oversight, this member was never assigned to for imported methods, and neither was BaseMethod. If class A exported methods x() and x(t), and B overrode A.x(), it would be impossible to call A.x(t) through a known instance of B. This fixes that problem.