Remove 'stack' parameter from __extern bodies
The 'stack' parameter was never needed. The only reason it exists for regular methods is so the method initializer can pre-allocate stack slots during method analysis. Native methods are not initialized in that way, and the evaluation stack grows dynamically anyway, so 'stack' is unnecessary. Plus it would be a nightmare to have to maintain...