BREAKING: Change use directive syntax
The module import directive, 'use x;', is removed. Only '/import' can be used to import modules now. The purpose of this is to pave the way for better dependency management, in the shape of build files, which will be implemented in ospc (https://github.com/osprey-lang/ospc).

At the same time, the namespace import directive, 'use namespace x;', is shortened to just 'use x;', which gets rid of one of Osprey's more annoying warts. The new syntax fits better into the language.

Needless to say, this change breaks an awful lot of code - in fact, just about all Osprey source files are now invalid! Oh dear.