BREAKING: Change lambda expression syntax
The form '@p = expr' has led to some confusion, with some people mistaking the '=' to mean assignment (as it does everywhere else). For that reason, the syntax is changed to '@p => expr'. The fat arrow is used in many other languages with a very similar meaning.

See http://osprey-lang.net/misc/lambda-syntax.txt for more